pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pydepsi"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "matplotlib",
    "sarxarray",
    "stmtools",
]
description = "Python package for interferometric synthetic aperture radar (InSAR) data processing."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Ou Ku", email = "o.ku@esciencecenter.nl"},
]
keywords = ["radar", "sar", "insar", "earth observation", "distributed computing"]
classifiers=[
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pycodestyle",
    "pre-commit",
    "ruff",
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-jupyter",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
]
demo = [
    "jupyterlab",
]

[tool.setuptools]
packages = ["pydepsi"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report term"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
source = ["pydepsi"]

[tool.ruff]
lint.select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "B",  # flake8-bugbear
  "D",  # pydocstyle
  "I",  # isort
  "N",  # PEP8-naming
  "UP",  # pyupgrade (upgrade syntax to current syntax)
  "PLE",  # Pylint error https://github.com/charliermarsh/ruff#error-ple
]
lint.ignore = [
"D100", "D101", "D104", "D105", "D106", "D107", "D203", "D213", "D413"
] # docstring style

line-length = 120
exclude = ["docs", "build", "*.ipynb", "examples/*"]
# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py310"

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]
